refactor(api): move db schema and seed to src/database
