+++ apps/api/src/database/seed/data-static.js
@@ -3,8 +3,8 @@
  */
 
 /**
- * @typedef {import('../src/server/applications/application/application.js')} Sector
- * @typedef {import('../src/server/applications/application/application.js').Sector} SubSector
+ * @typedef {import('../../server/applications/application/application.js')} Sector
+ * @typedef {import('../../server/applications/application/application.js').Sector} SubSector
  * @typedef {import('apps/web/src/server/applications/applications.types').Region} Region
  * @typedef {import('apps/web/src/server/applications/applications.types').ZoomLevel} ZoomLevel
  * @typedef {import('apps/web/src/server/applications/applications.types').ExaminationTimetableType} ExaminationTimetableType
